+++ ethereum/p2p/src/main/java/org/hyperledger/besu/ethereum/p2p/discovery/P2PDiscoveryConfiguration.java
@@ -28,6 +28,7 @@ public record P2PDiscoveryConfiguration(
     String p2pHost,
     String p2pInterface,
     Integer p2pPort,
+    Boolean p2pHostShareViaRlpxEnabled,
     Integer maxPeers,
     Boolean isLimitRemoteWireConnectionsEnabled,
     Percentage maxRemoteConnectionsPercentage,